Moved new layer methods to Group. Added more unit tests to Layer.
