+++ Source/Group.js
@@ -7,6 +7,7 @@
 
 import { Layer } from './Layer.js'
 import { Rectangle } from './Rectangle.js'
+import { Style } from './Style.js'
 
 
 /**
@@ -97,6 +98,143 @@ export class Group extends Layer {
   }
 
 
+
+
+    /**
+    Add a new wrapped layer object to represent a Sketch layer.
+    Apply any supplied properties to the wrapper (which will apply
+    them in turn to the wrapped layer).
+
+    @param {MSLayer} newLayer The underlying Sketch layer object.
+    @param {dictionary} properties The properties to apply.
+    @param {string} wrapper The name of wrapper class to use.
+    @return {Layer} The wrapped layer object.
+    */
+
+    _addWrappedLayerWithProperties(newLayer, properties, wrapper) {
+      if (newLayer) {
+        // add the Sketch object to this layer
+        var layer = this._object
+        layer.addLayers_(NSArray.arrayWithObject_(newLayer))
+
+        // make a Javascript wrapper object for the new layer
+        var wrapper = this._document.wrapObject(newLayer)
+
+        // apply properties, via the wrapper
+        for (var p in properties) {
+          wrapper[p] = properties[p]
+        }
+
+        return wrapper
+      }
+    }
+
+    /**
+    Extract the frame to use for a layer from some properties.
+    If the frame wasn't supplied in the properties, we return a default value instead.
+
+    @param {dictionary} properties The properties to use when looking for a frame value.
+    @return {Rectangle} The frame rectangle to use.
+    */
+
+    _frameForLayerWithProperties(properties) {
+      var frame = properties.frame
+      if (frame) {
+        delete properties["frame"]
+      } else {
+        frame = new Rectangle(0, 0, 100, 100)
+      }
+      return frame
+    }
+
+    /**
+    Extract the style to use for a layer from some properties.
+    If the style wasn't supplied at all, we use the default one.
+    */
+
+    _styleForLayerWithProperties(properties) {
+      var style = properties.style
+      if (!style) {
+        style = new Style()
+      }
+
+      var fills = properties.fills
+      if (fills) {
+        delete properties["fills"]
+        style.fills = fills
+      }
+
+      var borders = properties.borders
+      if (borders) {
+        delete properties["borders"]
+        style.borders = borders
+      }
+
+      return style
+    }
+
+    /**
+    Returns a newly created shape, which has been added to this layer,
+    and sets it up using the supplied properties.
+
+    @param {dictionary} properties Properties to apply to the shape.
+    @return {Shape} the new shape.
+    */
+
+    newShape(properties = {}) {
+      var frame = this._frameForLayerWithProperties(properties)
+      var newLayer = MSShapeGroup.shapeWithBezierPath_(NSBezierPath.bezierPathWithRect_(frame.asCGRect()));
+      properties["style"] = this._styleForLayerWithProperties(properties)
+
+      return this._addWrappedLayerWithProperties(newLayer, properties, "Shape");
+    }
+
+    /**
+    Returns a newly created text layer, which has been added to this layer,
+    and sets it up using the supplied properties.
+
+    @param {dictionary} properties Properties to apply to the text layer.
+    @return {Text} the new text layer.
+    */
+
+    newText(properties = {}) {
+      var frame = this._frameForLayerWithProperties(properties)
+      var newLayer = MSTextLayer.alloc().initWithFrame_(frame.asCGRect());
+      newLayer.adjustFrameToFit();
+      return this._addWrappedLayerWithProperties(newLayer, properties, "Text");
+    }
+
+    /**
+    Returns a newly created group, which has been added to this layer,
+    and sets it up using the supplied properties.
+
+    @param {dictionary} properties Properties to apply to the group.
+    @return {Group} the new group.
+    */
+
+    newGroup(properties = {}) {
+      var frame = this._frameForLayerWithProperties(properties)
+      var newLayer = MSLayerGroup.alloc().initWithFrame_(frame.asCGRect());
+      return this._addWrappedLayerWithProperties(newLayer, properties, "Group");
+    }
+
+
+    /**
+    Returns a newly created image layer, which has been added to this layer,
+    and sets it up using the supplied properties.
+
+    @param {dictionary} properties Properties to apply to the layer.
+    @return {Image} the new image layer.
+    */
+
+    newImage(properties = {}) {
+      var frame = this._frameForLayerWithProperties(properties)
+      var newLayer = MSBitmapLayer.alloc().initWithFrame_(frame.asCGRect());
+      return this._addWrappedLayerWithProperties(newLayer, properties, "Image");
+    }
+
+
+
   /**
   Return a list of tests to run for this class.
 
@@ -143,10 +281,7 @@ export class Group extends Layer {
           var page = document.selectedPage
           var group = page.newGroup({"frame" : new Rectangle(100, 100, 100, 100)})
           var local = group.pageRectToLocalRect(new Rectangle(125, 75, 50, 200))
-          tester.assertEqual(local.x, 25)
-          tester.assertEqual(local.y, -25)
-          tester.assertEqual(local.width, 50)
-          tester.assertEqual(local.height, 200)
+          tester.assertEqual(local, new Rectangle(25, -25, 50, 200))
         },
 
         "testAdjustToFit" : function(tester) {
@@ -156,10 +291,7 @@ export class Group extends Layer {
           var text = group.newShape({"frame" : new Rectangle(50, 50, 50, 50)})
           group.adjustToFit()
           var frame = group.frame
-          tester.assertEqual(frame.x, 150)
-          tester.assertEqual(frame.y, 150)
-          tester.assertEqual(frame.width, 50)
-          tester.assertEqual(frame.height, 50)
+          tester.assertEqual(frame, new Rectangle(150, 150, 50, 50))
         },
 
         "testIsGroup" : function(tester) {
